ai_service/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0